renamed row to board. Because that's what it is
